Refactor Profile Page and Components
- Updated Profile component to use new ProfileHeader, ProfileButton, and ProfileInfoItem components for better structure and readability.
- Enhanced ProfileHeader to display user cover image and avatar with improved styling.
- Created ProfileButton component for consistent button styling in profile actions.
- Implemented ProfileInfoItem component to display user information with icons.
- Added validation for user token in API routes to return appropriate error messages.
- Improved Sidebar component to dynamically render sections based on the current path.
- Updated Footer component to conditionally render based on the current route.
- Added Note page with functionality to handle multiple questions and mock processing of answers.
- Enhanced Navbar to include new navigation items and improved user authentication handling.
- Updated global types to include SectionItem and SectionMap interfaces for better type safety.

+++ src/app/api/user/me/route.ts
@@ -7,6 +7,9 @@ import { sendEmail } from "@/helper/mail/mailer";
 export async function GET(req: NextRequest) {
   try {
     const userId = getDataFromToken(req);
+    if(!userId){
+      return NextResponse.json({message:"Invalid Token"},{status:401});
+    }
     const user=await prisma.user.findFirst({
       where:{id: userId},
       omit:{
@@ -34,6 +37,9 @@ export async function GET(req: NextRequest) {
 export async function POST(req: NextRequest) {
   try {
     const userId = getDataFromToken(req);
+    if(!userId){
+      return NextResponse.json({message:"Invalid Token"},{status:401});
+    }
     const {basicInfo,updateEmail,updatePassword}=await req.json();
     let updatedUser;
     if(basicInfo){
@@ -91,6 +97,9 @@ export async function POST(req: NextRequest) {
 export async function DELETE(req: NextRequest) {
   try {
     const userId = getDataFromToken(req);
+    if(!userId){
+      return NextResponse.json({message:"Invalid Token"},{status:401});
+    }
     await prisma.user.delete({
       where:{id:userId}
     });
